Correcciones en la parte 1.

+++ 2025_TI301_MPTD_MAFT/src/unidad1/proglineal/NumeroOptTest.java
@@ -4,7 +4,7 @@
  * Docente: Francisco Montecillo Puente 
  * Estudiante / Programador: Angel Flores
  * Descripción: Prueba de NumerOpt
- *              i. e. [1, -3, -6, 9, 0, 1, -2]; R: -6
+ *              i. e. [1, -3, -6, 9, 0, 1, -2]; R: 2
  *              i. e. [1, 2, 3, 4]; R: -1
  * E-mail: miguelangelflto6@gmail.com
  * E-mail institucional: ti24110044@itess.edu.mx
@@ -17,7 +17,7 @@ public class NumeroOptTest {
     public static void main(String...args) {
         // Test 1
         System.out.println("Test número 1");
-        NumeroOpt numOpt = new NumeroOpt(new int[] {1, -3, -6, 9, 0, 1, -2});
+        NumeroOpt numOpt = new NumeroOpt(new double[] {1, -3, -6, 9, 0, 1, -2});
         
         // Test Obj1 resolver()
         System.out.println("resolver(): " + numOpt.resolver());
@@ -29,7 +29,7 @@ public class NumeroOptTest {
         //Test 2
         System.out.println("\nTest número 2");
         
-        numOpt = new NumeroOpt(new int[] {1, 2, 3, 4});
+        numOpt = new NumeroOpt(new double[] {1, 2, 3, 4});
         
         // Test obj2 resolver()
         System.out.println("resolver(): " + numOpt.resolver());
